fastapi-backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
tensorflow
pillow
numpy
python-multipart
bcrypt<4.0.0
